Refactor common schema code into SchemaLoader
