+++ src/ClientGenerator.cpp
@@ -48,53 +48,17 @@
 
 using namespace std::literals;
 
-namespace graphql::client {
+namespace graphql::generator::client {
 
-using namespace generator;
-
-const BuiltinTypeMap Generator::s_builtinTypes = {
-	{ "Int"sv, BuiltinType::Int },
-	{ "Float"sv, BuiltinType::Float },
-	{ "String"sv, BuiltinType::String },
-	{ "Boolean"sv, BuiltinType::Boolean },
-	{ "ID"sv, BuiltinType::ID },
-};
-
-const CppTypeMap Generator::s_builtinCppTypes = {
-	"response::IntType"sv,
-	"response::FloatType"sv,
-	"response::StringType"sv,
-	"response::BooleanType"sv,
-	"response::IdType"sv,
-};
-
-const std::string_view Generator::s_scalarCppType = R"cpp(response::Value)cpp"sv;
-
-Generator::Generator(GeneratorOptions&& options)
-	: _options(std::move(options))
-	, _clientNamespace(_options.client.clientNamespace)
+Generator::Generator(SchemaOptions&& schemaOptions, GeneratorOptions&& options)
+	: _loader(std::make_optional(std::move(schemaOptions)))
+	, _options(std::move(options))
 	, _headerDir(getHeaderDir())
 	, _sourceDir(getSourceDir())
 	, _headerPath(getHeaderPath())
 	, _sourcePath(getSourcePath())
 {
-
-	_schema = peg::parseSchemaFile(_options.client.schemaFilename);
-
-	if (!_schema.root)
-	{
-		throw std::logic_error("Unable to parse the service schema, but there was no error "
-							   "message from the parser!");
-	}
-
-	for (const auto& child : _schema.root->children)
-	{
-		visitDefinition(*child);
-	}
-
-	validateSchema();
-
-	_request = peg::parseFile(_options.client.requestFilename);
+	_request = peg::parseFile(_options.requestFilename);
 
 	if (!_request.root)
 	{
@@ -131,7 +95,7 @@ std::string Generator::getHeaderPath() const noexcept
 {
 	fs::path fullPath { _headerDir };
 
-	fullPath /= (_options.client.filenamePrefix + "Client.h");
+	fullPath /= (std::string { _loader.getFilenamePrefix() } + "Client.h");
 
 	return fullPath.string();
 }
@@ -140,324 +104,9 @@ std::string Generator::getSourcePath() const noexcept
 {
 	fs::path fullPath { _sourceDir };
 
-	fullPath /= (_options.client.filenamePrefix + "Client.cpp");
+	fullPath /= (std::string { _loader.getFilenamePrefix() } + "Client.cpp");
 
 	return fullPath.string();
-}
-
-void Generator::validateSchema()
-{
-	// Verify that none of the custom types conflict with a built-in type.
-	for (const auto& entry : _clientTypes)
-	{
-		if (s_builtinTypes.find(entry.first) != s_builtinTypes.cend())
-		{
-			std::ostringstream error;
-			auto itrPosition = _typePositions.find(entry.first);
-
-			error << "Builtin type overridden: " << entry.first;
-
-			if (itrPosition != _typePositions.cend())
-			{
-				error << " line: " << itrPosition->second.line
-					  << " column: " << itrPosition->second.column;
-			}
-
-			throw std::runtime_error(error.str());
-		}
-	}
-
-	// Fixup all of the fieldType members.
-	for (auto& entry : _inputTypes)
-	{
-		fixupInputFieldList(entry.fields);
-	}
-
-	for (auto& entry : _interfaceTypes)
-	{
-		fixupOutputFieldList(entry.fields, std::nullopt, std::nullopt);
-	}
-
-	bool queryDefined = false;
-
-	if (_operationTypes.empty())
-	{
-		// Fill in the operations with default type names if present.
-		constexpr auto strDefaultQuery = "Query"sv;
-		constexpr auto strDefaultMutation = "Mutation"sv;
-		constexpr auto strDefaultSubscription = "Subscription"sv;
-
-		if (_objectNames.find(strDefaultQuery) != _objectNames.cend())
-		{
-			_operationTypes.push_back(
-				{ strDefaultQuery, getSafeCppName(strDefaultQuery), service::strQuery });
-			queryDefined = true;
-		}
-
-		if (_objectNames.find(strDefaultMutation) != _objectNames.cend())
-		{
-			_operationTypes.push_back(
-				{ strDefaultMutation, getSafeCppName(strDefaultMutation), service::strMutation });
-		}
-
-		if (_objectNames.find(strDefaultSubscription) != _objectNames.cend())
-		{
-			_operationTypes.push_back({ strDefaultSubscription,
-				getSafeCppName(strDefaultSubscription),
-				service::strSubscription });
-		}
-	}
-	else
-	{
-		// Validate that all of the operation types exist and that query is defined.
-		for (const auto& operation : _operationTypes)
-		{
-			if (_objectNames.find(operation.type) == _objectNames.cend())
-			{
-				std::ostringstream error;
-
-				error << "Unknown operation type: " << operation.type
-					  << " operation: " << operation.operation;
-
-				throw std::runtime_error(error.str());
-			}
-
-			queryDefined = queryDefined || (operation.operation == service::strQuery);
-		}
-	}
-
-	if (!queryDefined)
-	{
-		throw std::runtime_error("Query operation type undefined");
-	}
-
-	std::string_view mutationType;
-
-	for (const auto& operation : _operationTypes)
-	{
-		if (operation.operation == service::strMutation)
-		{
-			mutationType = operation.type;
-			break;
-		}
-	}
-
-	for (auto& entry : _objectTypes)
-	{
-		auto interfaceFields = std::make_optional<std::unordered_set<std::string_view>>();
-		auto accessor = (mutationType == entry.type)
-			? std::make_optional<std::string_view>(strApply)
-			: std::nullopt;
-
-		for (const auto& interfaceName : entry.interfaces)
-		{
-			auto itr = _interfaceNames.find(interfaceName);
-
-			if (itr != _interfaceNames.cend())
-			{
-				for (const auto& field : _interfaceTypes[itr->second].fields)
-				{
-					interfaceFields->insert(field.name);
-				}
-			}
-		}
-
-		fixupOutputFieldList(entry.fields, interfaceFields, accessor);
-	}
-
-	// Validate the interfaces implemented by the object types.
-	for (const auto& entry : _objectTypes)
-	{
-		for (const auto& interfaceName : entry.interfaces)
-		{
-			if (_interfaceNames.find(interfaceName) == _interfaceNames.cend())
-			{
-				std::ostringstream error;
-				auto itrPosition = _typePositions.find(entry.type);
-
-				error << "Unknown interface: " << interfaceName
-					  << " implemented by: " << entry.type;
-
-				if (itrPosition != _typePositions.cend())
-				{
-					error << " line: " << itrPosition->second.line
-						  << " column: " << itrPosition->second.column;
-				}
-
-				throw std::runtime_error(error.str());
-			}
-		}
-	}
-
-	// Validate the objects that are possible types for unions and add the unions to
-	// the list of matching types for the objects.
-	for (const auto& entry : _unionTypes)
-	{
-		for (const auto& objectName : entry.options)
-		{
-			auto itr = _objectNames.find(objectName);
-
-			if (itr == _objectNames.cend())
-			{
-				std::ostringstream error;
-				auto itrPosition = _typePositions.find(entry.type);
-
-				error << "Unknown type: " << objectName << " included by: " << entry.type;
-
-				if (itrPosition != _typePositions.cend())
-				{
-					error << " line: " << itrPosition->second.line
-						  << " column: " << itrPosition->second.column;
-				}
-
-				throw std::runtime_error(error.str());
-			}
-
-			_objectTypes[itr->second].unions.push_back(entry.type);
-		}
-	}
-}
-
-void Generator::fixupOutputFieldList(OutputFieldList& fields,
-	const std::optional<std::unordered_set<std::string_view>>& interfaceFields,
-	const std::optional<std::string_view>& accessor)
-{
-	for (auto& entry : fields)
-	{
-		if (interfaceFields)
-		{
-			entry.interfaceField = false;
-			entry.inheritedField = interfaceFields->find(entry.name) != interfaceFields->cend();
-		}
-		else
-		{
-			entry.interfaceField = true;
-			entry.inheritedField = false;
-		}
-
-		if (accessor)
-		{
-			entry.accessor = *accessor;
-		}
-
-		if (s_builtinTypes.find(entry.type) != s_builtinTypes.cend())
-		{
-			continue;
-		}
-
-		auto itr = _clientTypes.find(entry.type);
-
-		if (itr == _clientTypes.cend())
-		{
-			std::ostringstream error;
-
-			error << "Unknown field type: " << entry.type;
-
-			if (entry.position)
-			{
-				error << " line: " << entry.position->line << " column: " << entry.position->column;
-			}
-
-			throw std::runtime_error(error.str());
-		}
-
-		switch (itr->second)
-		{
-			case SchemaType::Scalar:
-				entry.fieldType = OutputFieldType::Scalar;
-				break;
-
-			case SchemaType::Enum:
-				entry.fieldType = OutputFieldType::Enum;
-				break;
-
-			case SchemaType::Union:
-				entry.fieldType = OutputFieldType::Union;
-				break;
-
-			case SchemaType::Interface:
-				entry.fieldType = OutputFieldType::Interface;
-				break;
-
-			case SchemaType::Object:
-				entry.fieldType = OutputFieldType::Object;
-				break;
-
-			default:
-			{
-				std::ostringstream error;
-
-				error << "Invalid field type: " << entry.type;
-
-				if (entry.position)
-				{
-					error << " line: " << entry.position->line
-						  << " column: " << entry.position->column;
-				}
-
-				throw std::runtime_error(error.str());
-			}
-		}
-
-		fixupInputFieldList(entry.arguments);
-	}
-}
-
-void Generator::fixupInputFieldList(InputFieldList& fields)
-{
-	for (auto& entry : fields)
-	{
-		if (s_builtinTypes.find(entry.type) != s_builtinTypes.cend())
-		{
-			continue;
-		}
-
-		auto itr = _clientTypes.find(entry.type);
-
-		if (itr == _clientTypes.cend())
-		{
-			std::ostringstream error;
-
-			error << "Unknown argument type: " << entry.type;
-
-			if (entry.position)
-			{
-				error << " line: " << entry.position->line << " column: " << entry.position->column;
-			}
-
-			throw std::runtime_error(error.str());
-		}
-
-		switch (itr->second)
-		{
-			case SchemaType::Scalar:
-				entry.fieldType = InputFieldType::Scalar;
-				break;
-
-			case SchemaType::Enum:
-				entry.fieldType = InputFieldType::Enum;
-				break;
-
-			case SchemaType::Input:
-				entry.fieldType = InputFieldType::Input;
-				break;
-
-			default:
-			{
-				std::ostringstream error;
-
-				error << "Invalid argument type: " << entry.type;
-
-				if (entry.position)
-				{
-					error << " line: " << entry.position->line
-						  << " column: " << entry.position->column;
-				}
-
-				throw std::runtime_error(error.str());
-			}
-		}
-	}
 }
 
 void Generator::validateQuery() const
@@ -487,9 +136,9 @@ std::shared_ptr<schema::Schema> Generator::buildSchema() const
 
 void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) const
 {
-	if (!_scalarTypes.empty())
+	if (!_loader.getScalarTypes().empty())
 	{
-		for (const auto& scalarType : _scalarTypes)
+		for (const auto& scalarType : _loader.getScalarTypes())
 		{
 			schema->AddType(scalarType.type,
 				schema::ScalarType::Make(scalarType.type, scalarType.description));
@@ -498,9 +147,9 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 
 	std::map<std::string_view, std::shared_ptr<schema::EnumType>> enumTypes;
 
-	if (!_enumTypes.empty())
+	if (!_loader.getEnumTypes().empty())
 	{
-		for (const auto& enumType : _enumTypes)
+		for (const auto& enumType : _loader.getEnumTypes())
 		{
 			const auto itr = enumTypes
 								 .emplace(std::make_pair(enumType.type,
@@ -513,9 +162,9 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 
 	std::map<std::string_view, std::shared_ptr<schema::InputObjectType>> inputTypes;
 
-	if (!_inputTypes.empty())
+	if (!_loader.getInputTypes().empty())
 	{
-		for (const auto& inputType : _inputTypes)
+		for (const auto& inputType : _loader.getInputTypes())
 		{
 			const auto itr =
 				inputTypes
@@ -529,9 +178,9 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 
 	std::map<std::string_view, std::shared_ptr<schema::UnionType>> unionTypes;
 
-	if (!_unionTypes.empty())
+	if (!_loader.getUnionTypes().empty())
 	{
-		for (const auto& unionType : _unionTypes)
+		for (const auto& unionType : _loader.getUnionTypes())
 		{
 			const auto itr =
 				unionTypes
@@ -545,9 +194,9 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 
 	std::map<std::string_view, std::shared_ptr<schema::InterfaceType>> interfaceTypes;
 
-	if (!_interfaceTypes.empty())
+	if (!_loader.getInterfaceTypes().empty())
 	{
-		for (const auto& interfaceType : _interfaceTypes)
+		for (const auto& interfaceType : _loader.getInterfaceTypes())
 		{
 			const auto itr =
 				interfaceTypes
@@ -561,9 +210,9 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 
 	std::map<std::string_view, std::shared_ptr<schema::ObjectType>> objectTypes;
 
-	if (!_objectTypes.empty())
+	if (!_loader.getObjectTypes().empty())
 	{
-		for (const auto& objectType : _objectTypes)
+		for (const auto& objectType : _loader.getObjectTypes())
 		{
 			const auto itr =
 				objectTypes
@@ -575,7 +224,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 		}
 	}
 
-	for (const auto& enumType : _enumTypes)
+	for (const auto& enumType : _loader.getEnumTypes())
 	{
 		const auto itr = enumTypes.find(enumType.type);
 
@@ -598,7 +247,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 		}
 	}
 
-	for (const auto& inputType : _inputTypes)
+	for (const auto& inputType : _loader.getInputTypes())
 	{
 		const auto itr = inputTypes.find(inputType.type);
 
@@ -620,7 +269,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 		}
 	}
 
-	for (const auto& unionType : _unionTypes)
+	for (const auto& unionType : _loader.getUnionTypes())
 	{
 		const auto itr = unionTypes.find(unionType.type);
 
@@ -639,7 +288,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 		}
 	}
 
-	for (const auto& interfaceType : _interfaceTypes)
+	for (const auto& interfaceType : _loader.getInterfaceTypes())
 	{
 		const auto itr = interfaceTypes.find(interfaceType.type);
 
@@ -675,7 +324,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 		}
 	}
 
-	for (const auto& objectType : _objectTypes)
+	for (const auto& objectType : _loader.getObjectTypes())
 	{
 		const auto itr = objectTypes.find(objectType.type);
 
@@ -726,7 +375,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 		}
 	}
 
-	for (const auto& directive : _directives)
+	for (const auto& directive : _loader.getDirectives())
 	{
 		std::vector<introspection::DirectiveLocation> locations(directive.locations.size());
 
@@ -761,7 +410,7 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 			std::move(arguments)));
 	}
 
-	for (const auto& operationType : _operationTypes)
+	for (const auto& operationType : _loader.getOperationTypes())
 	{
 		const auto itr = objectTypes.find(operationType.type);
 
@@ -780,818 +429,6 @@ void Generator::addTypesToSchema(const std::shared_ptr<schema::Schema>& schema) 
 	}
 }
 
-void Generator::visitDefinition(const peg::ast_node& definition)
-{
-	if (definition.is_type<peg::schema_definition>())
-	{
-		visitClientDefinition(definition);
-	}
-	else if (definition.is_type<peg::schema_extension>())
-	{
-		visitClientExtension(definition);
-	}
-	else if (definition.is_type<peg::scalar_type_definition>())
-	{
-		visitScalarTypeDefinition(definition);
-	}
-	else if (definition.is_type<peg::enum_type_definition>())
-	{
-		visitEnumTypeDefinition(definition);
-	}
-	else if (definition.is_type<peg::enum_type_extension>())
-	{
-		visitEnumTypeExtension(definition);
-	}
-	else if (definition.is_type<peg::input_object_type_definition>())
-	{
-		visitInputObjectTypeDefinition(definition);
-	}
-	else if (definition.is_type<peg::input_object_type_extension>())
-	{
-		visitInputObjectTypeExtension(definition);
-	}
-	else if (definition.is_type<peg::union_type_definition>())
-	{
-		visitUnionTypeDefinition(definition);
-	}
-	else if (definition.is_type<peg::union_type_extension>())
-	{
-		visitUnionTypeExtension(definition);
-	}
-	else if (definition.is_type<peg::interface_type_definition>())
-	{
-		visitInterfaceTypeDefinition(definition);
-	}
-	else if (definition.is_type<peg::interface_type_extension>())
-	{
-		visitInterfaceTypeExtension(definition);
-	}
-	else if (definition.is_type<peg::object_type_definition>())
-	{
-		visitObjectTypeDefinition(definition);
-	}
-	else if (definition.is_type<peg::object_type_extension>())
-	{
-		visitObjectTypeExtension(definition);
-	}
-	else if (definition.is_type<peg::directive_definition>())
-	{
-		visitDirectiveDefinition(definition);
-	}
-	else
-	{
-		const auto position = definition.begin();
-		std::ostringstream error;
-
-		error << "Unexpected executable definition line: " << position.line
-			  << " column: " << position.column;
-
-		throw std::runtime_error(error.str());
-	}
-}
-
-void Generator::visitClientDefinition(const peg::ast_node& clientDefinition)
-{
-	peg::for_each_child<peg::root_operation_definition>(clientDefinition,
-		[this](const peg::ast_node& child) {
-			const auto operation(child.children.front()->string_view());
-			const auto name(child.children.back()->string_view());
-			const auto cppName(getSafeCppName(name));
-
-			_operationTypes.push_back({ name, cppName, operation });
-		});
-}
-
-void Generator::visitClientExtension(const peg::ast_node& clientExtension)
-{
-	peg::for_each_child<peg::operation_type_definition>(clientExtension,
-		[this](const peg::ast_node& child) {
-			const auto operation(child.children.front()->string_view());
-			const auto name(child.children.back()->string_view());
-			const auto cppName(getSafeCppName(name));
-
-			_operationTypes.push_back({ name, cppName, operation });
-		});
-}
-
-void Generator::visitObjectTypeDefinition(const peg::ast_node& objectTypeDefinition)
-{
-	std::string_view name;
-	std::string_view description;
-
-	peg::on_first_child<peg::object_name>(objectTypeDefinition,
-		[&name](const peg::ast_node& child) {
-			name = child.string_view();
-		});
-
-	peg::on_first_child<peg::description>(objectTypeDefinition,
-		[&description](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				description = child.children.front()->unescaped_view();
-			}
-		});
-
-	_clientTypes[name] = SchemaType::Object;
-	_typePositions.emplace(name, objectTypeDefinition.begin());
-	_objectNames[name] = _objectTypes.size();
-
-	auto cppName = getSafeCppName(name);
-
-	_objectTypes.push_back({ name, cppName, {}, {}, {}, description });
-
-	visitObjectTypeExtension(objectTypeDefinition);
-}
-
-void Generator::visitObjectTypeExtension(const peg::ast_node& objectTypeExtension)
-{
-	std::string_view name;
-
-	peg::on_first_child<peg::object_name>(objectTypeExtension, [&name](const peg::ast_node& child) {
-		name = child.string_view();
-	});
-
-	const auto itrType = _objectNames.find(name);
-
-	if (itrType != _objectNames.cend())
-	{
-		auto& objectType = _objectTypes[itrType->second];
-
-		peg::for_each_child<peg::interface_type>(objectTypeExtension,
-			[&objectType](const peg::ast_node& child) {
-				objectType.interfaces.push_back(child.string_view());
-			});
-
-		peg::on_first_child<peg::fields_definition>(objectTypeExtension,
-			[&objectType](const peg::ast_node& child) {
-				auto fields = getOutputFields(child.children);
-
-				objectType.fields.reserve(objectType.fields.size() + fields.size());
-				for (auto& field : fields)
-				{
-					objectType.fields.push_back(std::move(field));
-				}
-			});
-	}
-}
-
-void Generator::visitInterfaceTypeDefinition(const peg::ast_node& interfaceTypeDefinition)
-{
-	std::string_view name;
-	std::string_view description;
-
-	peg::on_first_child<peg::interface_name>(interfaceTypeDefinition,
-		[&name](const peg::ast_node& child) {
-			name = child.string_view();
-		});
-
-	peg::on_first_child<peg::description>(interfaceTypeDefinition,
-		[&description](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				description = child.children.front()->unescaped_view();
-			}
-		});
-
-	_clientTypes[name] = SchemaType::Interface;
-	_typePositions.emplace(name, interfaceTypeDefinition.begin());
-	_interfaceNames[name] = _interfaceTypes.size();
-
-	auto cppName = getSafeCppName(name);
-
-	_interfaceTypes.push_back({ name, cppName, {}, description });
-
-	visitInterfaceTypeExtension(interfaceTypeDefinition);
-}
-
-void Generator::visitInterfaceTypeExtension(const peg::ast_node& interfaceTypeExtension)
-{
-	std::string_view name;
-
-	peg::on_first_child<peg::interface_name>(interfaceTypeExtension,
-		[&name](const peg::ast_node& child) {
-			name = child.string_view();
-		});
-
-	const auto itrType = _interfaceNames.find(name);
-
-	if (itrType != _interfaceNames.cend())
-	{
-		auto& interfaceType = _interfaceTypes[itrType->second];
-
-		peg::on_first_child<peg::fields_definition>(interfaceTypeExtension,
-			[&interfaceType](const peg::ast_node& child) {
-				auto fields = getOutputFields(child.children);
-
-				interfaceType.fields.reserve(interfaceType.fields.size() + fields.size());
-				for (auto& field : fields)
-				{
-					interfaceType.fields.push_back(std::move(field));
-				}
-			});
-	}
-}
-
-void Generator::visitInputObjectTypeDefinition(const peg::ast_node& inputObjectTypeDefinition)
-{
-	std::string_view name;
-	std::string_view description;
-
-	peg::on_first_child<peg::object_name>(inputObjectTypeDefinition,
-		[&name](const peg::ast_node& child) {
-			name = child.string_view();
-		});
-
-	peg::on_first_child<peg::description>(inputObjectTypeDefinition,
-		[&description](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				description = child.children.front()->unescaped_view();
-			}
-		});
-
-	_clientTypes[name] = SchemaType::Input;
-	_typePositions.emplace(name, inputObjectTypeDefinition.begin());
-	_inputNames[name] = _inputTypes.size();
-
-	auto cppName = getSafeCppName(name);
-
-	_inputTypes.push_back({ name, cppName, {}, description });
-
-	visitInputObjectTypeExtension(inputObjectTypeDefinition);
-}
-
-void Generator::visitInputObjectTypeExtension(const peg::ast_node& inputObjectTypeExtension)
-{
-	std::string_view name;
-
-	peg::on_first_child<peg::object_name>(inputObjectTypeExtension,
-		[&name](const peg::ast_node& child) {
-			name = child.string_view();
-		});
-
-	const auto itrType = _inputNames.find(name);
-
-	if (itrType != _inputNames.cend())
-	{
-		auto& inputType = _inputTypes[itrType->second];
-
-		peg::on_first_child<peg::input_fields_definition>(inputObjectTypeExtension,
-			[&inputType](const peg::ast_node& child) {
-				auto fields = getInputFields(child.children);
-
-				inputType.fields.reserve(inputType.fields.size() + fields.size());
-				for (auto& field : fields)
-				{
-					inputType.fields.push_back(std::move(field));
-				}
-			});
-	}
-}
-
-void Generator::visitEnumTypeDefinition(const peg::ast_node& enumTypeDefinition)
-{
-	std::string_view name;
-	std::string_view description;
-
-	peg::on_first_child<peg::enum_name>(enumTypeDefinition, [&name](const peg::ast_node& child) {
-		name = child.string_view();
-	});
-
-	peg::on_first_child<peg::description>(enumTypeDefinition,
-		[&description](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				description = child.children.front()->unescaped_view();
-			}
-		});
-
-	_clientTypes[name] = SchemaType::Enum;
-	_typePositions.emplace(name, enumTypeDefinition.begin());
-	_enumNames[name] = _enumTypes.size();
-
-	auto cppName = getSafeCppName(name);
-
-	_enumTypes.push_back({ name, cppName, {}, description });
-
-	visitEnumTypeExtension(enumTypeDefinition);
-}
-
-void Generator::visitEnumTypeExtension(const peg::ast_node& enumTypeExtension)
-{
-	std::string_view name;
-
-	peg::on_first_child<peg::enum_name>(enumTypeExtension, [&name](const peg::ast_node& child) {
-		name = child.string_view();
-	});
-
-	const auto itrType = _enumNames.find(name);
-
-	if (itrType != _enumNames.cend())
-	{
-		auto& enumType = _enumTypes[itrType->second];
-
-		peg::for_each_child<
-			peg::enum_value_definition>(enumTypeExtension, [&enumType](const peg::ast_node& child) {
-			EnumValueType value;
-
-			peg::on_first_child<peg::enum_value>(child, [&value](const peg::ast_node& enumValue) {
-				value.value = enumValue.string_view();
-				value.cppValue = getSafeCppName(value.value);
-			});
-
-			peg::on_first_child<peg::description>(child,
-				[&value](const peg::ast_node& description) {
-					if (!description.children.empty())
-					{
-						value.description = description.children.front()->unescaped_view();
-					}
-				});
-
-			peg::on_first_child<peg::directives>(child, [&value](const peg::ast_node& directives) {
-				peg::for_each_child<peg::directive>(directives,
-					[&value](const peg::ast_node& directive) {
-						std::string_view directiveName;
-
-						peg::on_first_child<peg::directive_name>(directive,
-							[&directiveName](const peg::ast_node& name) {
-								directiveName = name.string_view();
-							});
-
-						if (directiveName == "deprecated")
-						{
-							std::string_view reason;
-
-							peg::on_first_child<peg::arguments>(directive,
-								[&value](const peg::ast_node& arguments) {
-									peg::on_first_child<peg::argument>(arguments,
-										[&value](const peg::ast_node& argument) {
-											std::string_view argumentName;
-
-											peg::on_first_child<peg::argument_name>(argument,
-												[&argumentName](const peg::ast_node& name) {
-													argumentName = name.string_view();
-												});
-
-											if (argumentName == "reason")
-											{
-												peg::on_first_child<peg::string_value>(argument,
-													[&value](const peg::ast_node& argumentValue) {
-														value.deprecationReason =
-															argumentValue.unescaped_view();
-													});
-											}
-										});
-								});
-						}
-					});
-			});
-
-			value.position = child.begin();
-			enumType.values.push_back(std::move(value));
-		});
-	}
-}
-
-void Generator::visitScalarTypeDefinition(const peg::ast_node& scalarTypeDefinition)
-{
-	std::string_view name;
-	std::string_view description;
-
-	peg::on_first_child<peg::scalar_name>(scalarTypeDefinition,
-		[&name](const peg::ast_node& child) {
-			name = child.string_view();
-		});
-
-	peg::on_first_child<peg::description>(scalarTypeDefinition,
-		[&description](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				description = child.children.front()->unescaped_view();
-			}
-		});
-
-	_clientTypes[name] = SchemaType::Scalar;
-	_typePositions.emplace(name, scalarTypeDefinition.begin());
-	_scalarNames[name] = _scalarTypes.size();
-	_scalarTypes.push_back({ name, description });
-}
-
-void Generator::visitUnionTypeDefinition(const peg::ast_node& unionTypeDefinition)
-{
-	std::string_view name;
-	std::string_view description;
-
-	peg::on_first_child<peg::union_name>(unionTypeDefinition, [&name](const peg::ast_node& child) {
-		name = child.string_view();
-	});
-
-	peg::on_first_child<peg::description>(unionTypeDefinition,
-		[&description](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				description = child.children.front()->unescaped_view();
-			}
-		});
-
-	_clientTypes[name] = SchemaType::Union;
-	_typePositions.emplace(name, unionTypeDefinition.begin());
-	_unionNames[name] = _unionTypes.size();
-
-	auto cppName = getSafeCppName(name);
-
-	_unionTypes.push_back({ name, cppName, {}, description });
-
-	visitUnionTypeExtension(unionTypeDefinition);
-}
-
-void Generator::visitUnionTypeExtension(const peg::ast_node& unionTypeExtension)
-{
-	std::string_view name;
-
-	peg::on_first_child<peg::union_name>(unionTypeExtension, [&name](const peg::ast_node& child) {
-		name = child.string_view();
-	});
-
-	const auto itrType = _unionNames.find(name);
-
-	if (itrType != _unionNames.cend())
-	{
-		auto& unionType = _unionTypes[itrType->second];
-
-		peg::for_each_child<peg::union_type>(unionTypeExtension,
-			[&unionType](const peg::ast_node& child) {
-				unionType.options.push_back(child.string_view());
-			});
-	}
-}
-
-void Generator::visitDirectiveDefinition(const peg::ast_node& directiveDefinition)
-{
-	Directive directive;
-
-	peg::on_first_child<peg::directive_name>(directiveDefinition,
-		[&directive](const peg::ast_node& child) {
-			directive.name = child.string_view();
-		});
-
-	peg::on_first_child<peg::description>(directiveDefinition,
-		[&directive](const peg::ast_node& child) {
-			if (!child.children.empty())
-			{
-				directive.description = child.children.front()->unescaped_view();
-			}
-		});
-
-	peg::for_each_child<peg::directive_location>(directiveDefinition,
-		[&directive](const peg::ast_node& child) {
-			directive.locations.push_back(child.string_view());
-		});
-
-	peg::on_first_child<peg::arguments_definition>(directiveDefinition,
-		[&directive](const peg::ast_node& child) {
-			auto fields = getInputFields(child.children);
-
-			directive.arguments.reserve(directive.arguments.size() + fields.size());
-			for (auto& field : fields)
-			{
-				directive.arguments.push_back(std::move(field));
-			}
-		});
-
-	_directivePositions.emplace(directive.name, directiveDefinition.begin());
-	_directives.push_back(std::move(directive));
-}
-
-std::string_view Generator::getSafeCppName(std::string_view type) noexcept
-{
-	// The C++ standard reserves all names starting with '_' followed by a capital letter,
-	// and all names that contain a double '_'. So we need to strip those from the types used
-	// in GraphQL when declaring C++ types.
-	static const std::regex multiple_(R"re(_{2,})re",
-		std::regex::optimize | std::regex::ECMAScript);
-	static const std::regex leading_Capital(R"re(^_([A-Z]))re",
-		std::regex::optimize | std::regex::ECMAScript);
-
-	// Cache the substitutions so we don't need to repeat a replacement.
-	static std::unordered_map<std::string_view, std::string> safeNames;
-	auto itr = safeNames.find(type);
-	std::string cppName { type };
-
-	if (safeNames.cend() == itr
-		&& (std::regex_search(cppName, leading_Capital) || std::regex_search(cppName, multiple_)))
-	{
-		std::tie(itr, std::ignore) = safeNames.emplace(type,
-			std::regex_replace(std::regex_replace(cppName, multiple_, R"re(_)re"),
-				leading_Capital,
-				R"re($1)re"));
-	}
-
-	return (safeNames.cend() == itr) ? type : itr->second;
-}
-
-OutputFieldList Generator::getOutputFields(const peg::ast_node::children_t& fields)
-{
-	OutputFieldList outputFields;
-
-	for (const auto& fieldDefinition : fields)
-	{
-		OutputField field;
-		TypeVisitor fieldType;
-
-		for (const auto& child : fieldDefinition->children)
-		{
-			if (child->is_type<peg::field_name>())
-			{
-				field.name = child->string_view();
-				field.cppName = getSafeCppName(field.name);
-			}
-			else if (child->is_type<peg::arguments_definition>())
-			{
-				field.arguments = getInputFields(child->children);
-			}
-			else if (child->is_type<peg::named_type>() || child->is_type<peg::list_type>()
-				|| child->is_type<peg::nonnull_type>())
-			{
-				fieldType.visit(*child);
-			}
-			else if (child->is_type<peg::description>() && !child->children.empty())
-			{
-				field.description = child->children.front()->unescaped_view();
-			}
-			else if (child->is_type<peg::directives>())
-			{
-				peg::for_each_child<peg::directive>(*child,
-					[&field](const peg::ast_node& directive) {
-						std::string_view directiveName;
-
-						peg::on_first_child<peg::directive_name>(directive,
-							[&directiveName](const peg::ast_node& name) {
-								directiveName = name.string_view();
-							});
-
-						if (directiveName == "deprecated")
-						{
-							std::string_view deprecationReason;
-
-							peg::on_first_child<peg::arguments>(directive,
-								[&deprecationReason](const peg::ast_node& arguments) {
-									peg::on_first_child<peg::argument>(arguments,
-										[&deprecationReason](const peg::ast_node& argument) {
-											std::string_view argumentName;
-
-											peg::on_first_child<peg::argument_name>(argument,
-												[&argumentName](const peg::ast_node& name) {
-													argumentName = name.string_view();
-												});
-
-											if (argumentName == "reason")
-											{
-												peg::on_first_child<peg::string_value>(argument,
-													[&deprecationReason](
-														const peg::ast_node& reason) {
-														deprecationReason = reason.unescaped_view();
-													});
-											}
-										});
-								});
-
-							field.deprecationReason = std::move(deprecationReason);
-						}
-					});
-			}
-		}
-
-		std::tie(field.type, field.modifiers) = fieldType.getType();
-		field.position = fieldDefinition->begin();
-		outputFields.push_back(std::move(field));
-	}
-
-	return outputFields;
-}
-
-InputFieldList Generator::getInputFields(const peg::ast_node::children_t& fields)
-{
-	InputFieldList inputFields;
-
-	for (const auto& fieldDefinition : fields)
-	{
-		InputField field;
-		TypeVisitor fieldType;
-		service::schema_location defaultValueLocation;
-
-		for (const auto& child : fieldDefinition->children)
-		{
-			if (child->is_type<peg::argument_name>())
-			{
-				field.name = child->string_view();
-				field.cppName = getSafeCppName(field.name);
-			}
-			else if (child->is_type<peg::named_type>() || child->is_type<peg::list_type>()
-				|| child->is_type<peg::nonnull_type>())
-			{
-				fieldType.visit(*child);
-			}
-			else if (child->is_type<peg::default_value>())
-			{
-				const auto position = child->begin();
-				DefaultValueVisitor defaultValue;
-
-				defaultValue.visit(*child->children.back());
-				field.defaultValue = defaultValue.getValue();
-				field.defaultValueString = child->children.back()->string_view();
-
-				defaultValueLocation = { position.line, position.column };
-			}
-			else if (child->is_type<peg::description>() && !child->children.empty())
-			{
-				field.description = child->children.front()->unescaped_view();
-			}
-		}
-
-		std::tie(field.type, field.modifiers) = fieldType.getType();
-		field.position = fieldDefinition->begin();
-
-		if (!field.defaultValueString.empty() && field.defaultValue.type() == response::Type::Null
-			&& (field.modifiers.empty()
-				|| field.modifiers.front() != service::TypeModifier::Nullable))
-		{
-			std::ostringstream error;
-
-			error << "Expected Non-Null default value for field name: " << field.name
-				  << " line: " << defaultValueLocation.line
-				  << " column: " << defaultValueLocation.column;
-
-			throw std::runtime_error(error.str());
-		}
-
-		inputFields.push_back(std::move(field));
-	}
-
-	return inputFields;
-}
-
-void Generator::TypeVisitor::visit(const peg::ast_node& typeName)
-{
-	if (typeName.is_type<peg::nonnull_type>())
-	{
-		visitNonNullType(typeName);
-	}
-	else if (typeName.is_type<peg::list_type>())
-	{
-		visitListType(typeName);
-	}
-	else if (typeName.is_type<peg::named_type>())
-	{
-		visitNamedType(typeName);
-	}
-}
-
-void Generator::TypeVisitor::visitNamedType(const peg::ast_node& namedType)
-{
-	if (!_nonNull)
-	{
-		_modifiers.push_back(service::TypeModifier::Nullable);
-	}
-
-	_type = namedType.string_view();
-}
-
-void Generator::TypeVisitor::visitListType(const peg::ast_node& listType)
-{
-	if (!_nonNull)
-	{
-		_modifiers.push_back(service::TypeModifier::Nullable);
-	}
-	_nonNull = false;
-
-	_modifiers.push_back(service::TypeModifier::List);
-
-	visit(*listType.children.front());
-}
-
-void Generator::TypeVisitor::visitNonNullType(const peg::ast_node& nonNullType)
-{
-	_nonNull = true;
-
-	visit(*nonNullType.children.front());
-}
-
-std::pair<std::string_view, TypeModifierStack> Generator::TypeVisitor::getType()
-{
-	return { std::move(_type), std::move(_modifiers) };
-}
-
-void Generator::DefaultValueVisitor::visit(const peg::ast_node& value)
-{
-	if (value.is_type<peg::integer_value>())
-	{
-		visitIntValue(value);
-	}
-	else if (value.is_type<peg::float_value>())
-	{
-		visitFloatValue(value);
-	}
-	else if (value.is_type<peg::string_value>())
-	{
-		visitStringValue(value);
-	}
-	else if (value.is_type<peg::true_keyword>() || value.is_type<peg::false_keyword>())
-	{
-		visitBooleanValue(value);
-	}
-	else if (value.is_type<peg::null_keyword>())
-	{
-		visitNullValue(value);
-	}
-	else if (value.is_type<peg::enum_value>())
-	{
-		visitEnumValue(value);
-	}
-	else if (value.is_type<peg::list_value>())
-	{
-		visitListValue(value);
-	}
-	else if (value.is_type<peg::object_value>())
-	{
-		visitObjectValue(value);
-	}
-	else if (value.is_type<peg::variable>())
-	{
-		std::ostringstream error;
-		const auto position = value.begin();
-
-		error << "Unexpected variable in default value line: " << position.line
-			  << " column: " << position.column;
-
-		throw std::runtime_error(error.str());
-	}
-}
-
-void Generator::DefaultValueVisitor::visitIntValue(const peg::ast_node& intValue)
-{
-	_value = response::Value(std::atoi(intValue.string().c_str()));
-}
-
-void Generator::DefaultValueVisitor::visitFloatValue(const peg::ast_node& floatValue)
-{
-	_value = response::Value(std::atof(floatValue.string().c_str()));
-}
-
-void Generator::DefaultValueVisitor::visitStringValue(const peg::ast_node& stringValue)
-{
-	_value = response::Value(std::string { stringValue.unescaped_view() });
-}
-
-void Generator::DefaultValueVisitor::visitBooleanValue(const peg::ast_node& booleanValue)
-{
-	_value = response::Value(booleanValue.is_type<peg::true_keyword>());
-}
-
-void Generator::DefaultValueVisitor::visitNullValue(const peg::ast_node& /*nullValue*/)
-{
-	_value = {};
-}
-
-void Generator::DefaultValueVisitor::visitEnumValue(const peg::ast_node& enumValue)
-{
-	_value = response::Value(response::Type::EnumValue);
-	_value.set<response::StringType>(enumValue.string());
-}
-
-void Generator::DefaultValueVisitor::visitListValue(const peg::ast_node& listValue)
-{
-	_value = response::Value(response::Type::List);
-	_value.reserve(listValue.children.size());
-
-	for (const auto& child : listValue.children)
-	{
-		DefaultValueVisitor visitor;
-
-		visitor.visit(*child);
-		_value.emplace_back(visitor.getValue());
-	}
-}
-
-void Generator::DefaultValueVisitor::visitObjectValue(const peg::ast_node& objectValue)
-{
-	_value = response::Value(response::Type::Map);
-	_value.reserve(objectValue.children.size());
-
-	for (const auto& field : objectValue.children)
-	{
-		DefaultValueVisitor visitor;
-
-		visitor.visit(*field->children.back());
-		_value.emplace_back(field->children.front()->string(), visitor.getValue());
-	}
-}
-
-response::Value Generator::DefaultValueVisitor::getValue()
-{
-	return response::Value(std::move(_value));
-}
-
 std::vector<std::string> Generator::Build() const noexcept
 {
 	std::vector<std::string> builtFiles;
@@ -1607,146 +444,6 @@ std::vector<std::string> Generator::Build() const noexcept
 	}
 
 	return builtFiles;
-}
-
-std::string_view Generator::getCppType(std::string_view type) const noexcept
-{
-	auto itrBuiltin = s_builtinTypes.find(type);
-
-	if (itrBuiltin != s_builtinTypes.cend())
-	{
-		if (static_cast<size_t>(itrBuiltin->second) < s_builtinCppTypes.size())
-		{
-			return s_builtinCppTypes[static_cast<size_t>(itrBuiltin->second)];
-		}
-	}
-	else
-	{
-		auto itrScalar = _scalarNames.find(type);
-
-		if (itrScalar != _scalarNames.cend())
-		{
-			return s_scalarCppType;
-		}
-	}
-
-	return getSafeCppName(type);
-}
-
-std::string Generator::getInputCppType(const InputField& field) const noexcept
-{
-	size_t templateCount = 0;
-	std::ostringstream inputType;
-
-	for (auto modifier : field.modifiers)
-	{
-		switch (modifier)
-		{
-			case service::TypeModifier::Nullable:
-				inputType << R"cpp(std::optional<)cpp";
-				++templateCount;
-				break;
-
-			case service::TypeModifier::List:
-				inputType << R"cpp(std::vector<)cpp";
-				++templateCount;
-				break;
-
-			default:
-				break;
-		}
-	}
-
-	inputType << getCppType(field.type);
-
-	for (size_t i = 0; i < templateCount; ++i)
-	{
-		inputType << R"cpp(>)cpp";
-	}
-
-	return inputType.str();
-}
-
-std::string Generator::getOutputCppType(const OutputField& field) const noexcept
-{
-	bool nonNull = true;
-	size_t templateCount = 0;
-	std::ostringstream outputType;
-
-	for (auto modifier : field.modifiers)
-	{
-		if (!nonNull)
-		{
-			outputType << R"cpp(std::optional<)cpp";
-			++templateCount;
-		}
-
-		switch (modifier)
-		{
-			case service::TypeModifier::None:
-				nonNull = true;
-				break;
-
-			case service::TypeModifier::Nullable:
-				nonNull = false;
-				break;
-
-			case service::TypeModifier::List:
-				nonNull = true;
-				outputType << R"cpp(std::vector<)cpp";
-				++templateCount;
-				break;
-		}
-	}
-
-	switch (field.fieldType)
-	{
-		case OutputFieldType::Object:
-		case OutputFieldType::Union:
-		case OutputFieldType::Interface:
-			// Even if it's non-nullable, we still want to return a shared_ptr for complex types
-			outputType << R"cpp(std::shared_ptr<)cpp";
-			++templateCount;
-			break;
-
-		default:
-			if (!nonNull)
-			{
-				outputType << R"cpp(std::optional<)cpp";
-				++templateCount;
-			}
-			break;
-	}
-
-	switch (field.fieldType)
-	{
-		case OutputFieldType::Builtin:
-		case OutputFieldType::Scalar:
-		case OutputFieldType::Enum:
-			outputType << getCppType(field.type);
-			break;
-
-		case OutputFieldType::Object:
-			if (field.interfaceField)
-			{
-				outputType << R"cpp(object::)cpp";
-			}
-
-			outputType << getSafeCppName(field.type);
-			break;
-
-		case OutputFieldType::Union:
-		case OutputFieldType::Interface:
-			outputType << R"cpp(service::Object)cpp";
-			break;
-	}
-
-	for (size_t i = 0; i < templateCount; ++i)
-	{
-		outputType << R"cpp(>)cpp";
-	}
-
-	return outputType.str();
 }
 
 bool Generator::outputHeader() const noexcept
@@ -1774,13 +471,13 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 )cpp";
 
 	NamespaceScope graphqlNamespace { headerFile, "graphql" };
-	NamespaceScope clientNamespace { headerFile, _clientNamespace };
+	NamespaceScope schemaNamespace { headerFile, _loader.getSchemaNamespace() };
 	NamespaceScope objectNamespace { headerFile, "object", true };
 	PendingBlankLine pendingSeparator { headerFile };
 
 	std::string_view queryType;
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		if (operation.operation == service::strQuery)
 		{
@@ -1789,11 +486,11 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 	}
 
-	if (!_enumTypes.empty())
+	if (!_loader.getEnumTypes().empty())
 	{
 		pendingSeparator.reset();
 
-		for (const auto& enumType : _enumTypes)
+		for (const auto& enumType : _loader.getEnumTypes())
 		{
 			headerFile << R"cpp(enum class )cpp" << enumType.cppType << R"cpp(
 {
@@ -1819,14 +516,14 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 	}
 
-	if (!_inputTypes.empty())
+	if (!_loader.getInputTypes().empty())
 	{
 		pendingSeparator.reset();
 
 		// Forward declare all of the input types
-		if (_inputTypes.size() > 1)
+		if (_loader.getInputTypes().size() > 1)
 		{
-			for (const auto& inputType : _inputTypes)
+			for (const auto& inputType : _loader.getInputTypes())
 			{
 				headerFile << R"cpp(struct )cpp" << inputType.cppType << R"cpp(;
 )cpp";
@@ -1836,7 +533,7 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 
 		// Output the full declarations
-		for (const auto& inputType : _inputTypes)
+		for (const auto& inputType : _loader.getInputTypes())
 		{
 			headerFile << R"cpp(struct )cpp" << inputType.cppType << R"cpp(
 {
@@ -1852,13 +549,13 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 	}
 
-	if (!_objectTypes.empty())
+	if (!_loader.getObjectTypes().empty())
 	{
 		objectNamespace.enter();
 		headerFile << std::endl;
 
 		// Forward declare all of the object types
-		for (const auto& objectType : _objectTypes)
+		for (const auto& objectType : _loader.getObjectTypes())
 		{
 			headerFile << R"cpp(class )cpp" << objectType.cppType << R"cpp(;
 )cpp";
@@ -1867,7 +564,7 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		headerFile << std::endl;
 	}
 
-	if (!_interfaceTypes.empty())
+	if (!_loader.getInterfaceTypes().empty())
 	{
 		if (objectNamespace.exit())
 		{
@@ -1875,9 +572,9 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 
 		// Forward declare all of the interface types
-		if (_interfaceTypes.size() > 1)
+		if (_loader.getInterfaceTypes().size() > 1)
 		{
-			for (const auto& interfaceType : _interfaceTypes)
+			for (const auto& interfaceType : _loader.getInterfaceTypes())
 			{
 				headerFile << R"cpp(struct )cpp" << interfaceType.cppType << R"cpp(;
 )cpp";
@@ -1887,7 +584,7 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 
 		// Output the full declarations
-		for (const auto& interfaceType : _interfaceTypes)
+		for (const auto& interfaceType : _loader.getInterfaceTypes())
 		{
 			headerFile << R"cpp(struct )cpp" << interfaceType.cppType << R"cpp(
 {
@@ -1904,7 +601,7 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 	}
 
-	if (!_objectTypes.empty())
+	if (!_loader.getObjectTypes().empty())
 	{
 		if (objectNamespace.enter())
 		{
@@ -1912,7 +609,7 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 		}
 
 		// Output the full declarations
-		for (const auto& objectType : _objectTypes)
+		for (const auto& objectType : _loader.getObjectTypes())
 		{
 			outputObjectDeclaration(headerFile, objectType, objectType.type == queryType);
 			headerFile << std::endl;
@@ -1932,7 +629,7 @@ static_assert(graphql::internal::MinorVersion == )cpp"
 public:
 	explicit Operations()cpp";
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		if (!firstOperation)
 		{
@@ -1949,7 +646,7 @@ public:
 private:
 )cpp";
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		headerFile << R"cpp(	std::shared_ptr<object::)cpp" << operation.cppType << R"cpp(> _)cpp"
 				   << operation.operation << R"cpp(;
@@ -1974,7 +671,7 @@ void Generator::outputObjectDeclaration(
 	for (const auto& interfaceName : objectType.interfaces)
 	{
 		headerFile << R"cpp(
-	, public )cpp" << getSafeCppName(interfaceName);
+	, public )cpp" << _loader.getSafeCppName(interfaceName);
 	}
 
 	headerFile << R"cpp(
@@ -2033,7 +730,7 @@ std::string Generator::getFieldDeclaration(const InputField& inputField) const n
 {
 	std::ostringstream output;
 
-	output << getInputCppType(inputField) << R"cpp( )cpp" << inputField.cppName;
+	output << _loader.getInputCppType(inputField) << R"cpp( )cpp" << inputField.cppName;
 
 	return output.str();
 }
@@ -2044,14 +741,14 @@ std::string Generator::getFieldDeclaration(const OutputField& outputField) const
 	std::string fieldName { outputField.cppName };
 
 	fieldName[0] = static_cast<char>(std::toupper(static_cast<unsigned char>(fieldName[0])));
-	output << R"cpp(	virtual service::FieldResult<)cpp" << getOutputCppType(outputField)
+	output << R"cpp(	virtual service::FieldResult<)cpp" << _loader.getOutputCppType(outputField)
 		   << R"cpp(> )cpp" << outputField.accessor << fieldName
 		   << R"cpp((service::FieldParams&& params)cpp";
 
 	for (const auto& argument : outputField.arguments)
 	{
-		output << R"cpp(, )cpp" << getInputCppType(argument) << R"cpp(&& )cpp" << argument.cppName
-			   << "Arg";
+		output << R"cpp(, )cpp" << _loader.getInputCppType(argument) << R"cpp(&& )cpp"
+			   << argument.cppName << "Arg";
 	}
 
 	output << R"cpp() const)cpp";
@@ -2110,13 +807,13 @@ using namespace std::literals;
 
 	NamespaceScope graphqlNamespace { sourceFile, "graphql" };
 
-	if (!_enumTypes.empty() || !_inputTypes.empty())
+	if (!_loader.getEnumTypes().empty() || !_loader.getInputTypes().empty())
 	{
 		NamespaceScope serviceNamespace { sourceFile, "service" };
 
 		sourceFile << std::endl;
 
-		for (const auto& enumType : _enumTypes)
+		for (const auto& enumType : _loader.getEnumTypes())
 		{
 			bool firstValue = true;
 
@@ -2141,9 +838,9 @@ using namespace std::literals;
 };
 
 template <>
-)cpp" << _clientNamespace
+)cpp" << _loader.getSchemaNamespace()
 					   << R"cpp(::)cpp" << enumType.cppType << R"cpp( ModifiedArgument<)cpp"
-					   << _clientNamespace << R"cpp(::)cpp" << enumType.cppType
+					   << _loader.getSchemaNamespace() << R"cpp(::)cpp" << enumType.cppType
 					   << R"cpp(>::convert(const response::Value& value)
 {
 	if (!value.maybe_enum())
@@ -2164,19 +861,19 @@ template <>
 	}
 
 	return static_cast<)cpp"
-					   << _clientNamespace << R"cpp(::)cpp" << enumType.cppType
+					   << _loader.getSchemaNamespace() << R"cpp(::)cpp" << enumType.cppType
 					   << R"cpp(>(itr - s_names)cpp" << enumType.cppType << R"cpp(.cbegin());
 }
 
 template <>
 std::future<service::ResolverResult> ModifiedResult<)cpp"
-					   << _clientNamespace << R"cpp(::)cpp" << enumType.cppType
-					   << R"cpp(>::convert(service::FieldResult<)cpp" << _clientNamespace
-					   << R"cpp(::)cpp" << enumType.cppType
+					   << _loader.getSchemaNamespace() << R"cpp(::)cpp" << enumType.cppType
+					   << R"cpp(>::convert(service::FieldResult<)cpp"
+					   << _loader.getSchemaNamespace() << R"cpp(::)cpp" << enumType.cppType
 					   << R"cpp(>&& result, ResolverParams&& params)
 {
 	return resolve(std::move(result), std::move(params),
-		[]()cpp" << _clientNamespace
+		[]()cpp" << _loader.getSchemaNamespace()
 					   << R"cpp(::)cpp" << enumType.cppType
 					   << R"cpp(&& value, const ResolverParams&)
 		{
@@ -2192,14 +889,14 @@ std::future<service::ResolverResult> ModifiedResult<)cpp"
 )cpp";
 		}
 
-		for (const auto& inputType : _inputTypes)
+		for (const auto& inputType : _loader.getInputTypes())
 		{
 			bool firstField = true;
 
 			sourceFile << R"cpp(template <>
-)cpp" << _clientNamespace
+)cpp" << _loader.getSchemaNamespace()
 					   << R"cpp(::)cpp" << inputType.cppType << R"cpp( ModifiedArgument<)cpp"
-					   << _clientNamespace << R"cpp(::)cpp" << inputType.cppType
+					   << _loader.getSchemaNamespace() << R"cpp(::)cpp" << inputType.cppType
 					   << R"cpp(>::convert(const response::Value& value)
 {
 )cpp";
@@ -2277,10 +974,10 @@ std::future<service::ResolverResult> ModifiedResult<)cpp"
 		sourceFile << std::endl;
 	}
 
-	NamespaceScope clientNamespace { sourceFile, _clientNamespace };
+	NamespaceScope schemaNamespace { sourceFile, _loader.getSchemaNamespace() };
 	std::string_view queryType;
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		if (operation.operation == service::strQuery)
 		{
@@ -2289,13 +986,13 @@ std::future<service::ResolverResult> ModifiedResult<)cpp"
 		}
 	}
 
-	if (!_objectTypes.empty())
+	if (!_loader.getObjectTypes().empty())
 	{
 		NamespaceScope objectNamespace { sourceFile, "object" };
 
 		sourceFile << std::endl;
 
-		for (const auto& objectType : _objectTypes)
+		for (const auto& objectType : _loader.getObjectTypes())
 		{
 			outputObjectImplementation(sourceFile, objectType, objectType.type == queryType);
 			sourceFile << std::endl;
@@ -2307,7 +1004,7 @@ std::future<service::ResolverResult> ModifiedResult<)cpp"
 	sourceFile << R"cpp(
 Operations::Operations()cpp";
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		if (!firstOperation)
 		{
@@ -2325,7 +1022,7 @@ Operations::Operations()cpp";
 
 	firstOperation = true;
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		if (!firstOperation)
 		{
@@ -2342,7 +1039,7 @@ Operations::Operations()cpp";
 	}, GetClient())
 )cpp";
 
-	for (const auto& operation : _operationTypes)
+	for (const auto& operation : _loader.getOperationTypes())
 	{
 		sourceFile << R"cpp(	, _)cpp" << operation.operation << R"cpp((std::move()cpp"
 				   << operation.operation << R"cpp())
@@ -2470,12 +1167,12 @@ void Generator::outputObjectImplementation(
 		fieldName[0] = static_cast<char>(std::toupper(static_cast<unsigned char>(fieldName[0])));
 		sourceFile << R"cpp(
 service::FieldResult<)cpp"
-				   << getOutputCppType(outputField) << R"cpp(> )cpp" << objectType.cppType
+				   << _loader.getOutputCppType(outputField) << R"cpp(> )cpp" << objectType.cppType
 				   << R"cpp(::)cpp" << outputField.accessor << fieldName
 				   << R"cpp((service::FieldParams&&)cpp";
 		for (const auto& argument : outputField.arguments)
 		{
-			sourceFile << R"cpp(, )cpp" << getInputCppType(argument) << R"cpp(&&)cpp";
+			sourceFile << R"cpp(, )cpp" << _loader.getInputCppType(argument) << R"cpp(&&)cpp";
 		}
 
 		sourceFile << R"cpp() const
@@ -2786,12 +1483,13 @@ std::string Generator::getArgumentAccessType(const InputField& argument) const n
 	switch (argument.fieldType)
 	{
 		case InputFieldType::Builtin:
-			argumentType << getCppType(argument.type);
+			argumentType << _loader.getCppType(argument.type);
 			break;
 
 		case InputFieldType::Enum:
 		case InputFieldType::Input:
-			argumentType << _clientNamespace << R"cpp(::)cpp" << getCppType(argument.type);
+			argumentType << _loader.getSchemaNamespace() << R"cpp(::)cpp"
+						 << _loader.getCppType(argument.type);
 			break;
 
 		case InputFieldType::Scalar:
@@ -2815,7 +1513,7 @@ std::string Generator::getResultAccessType(const OutputField& result) const noex
 		case OutputFieldType::Builtin:
 		case OutputFieldType::Enum:
 		case OutputFieldType::Object:
-			resultType << getCppType(result.type);
+			resultType << _loader.getCppType(result.type);
 			break;
 
 		case OutputFieldType::Scalar:
@@ -2909,7 +1607,7 @@ std::shared_ptr<const schema::BaseType> Generator::getIntrospectionType(
 	return introspectionType;
 }
 
-} /* namespace graphql::client */
+} /* namespace graphql::generator::client */
 
 namespace po = boost::program_options;
 
@@ -2939,7 +1637,7 @@ int main(int argc, char** argv)
 	std::string schemaFileName;
 	std::string requestFileName;
 	std::string filenamePrefix;
-	std::string clientNamespace;
+	std::string schemaNamespace;
 	std::string sourceDir;
 	std::string headerDir;
 
@@ -2955,7 +1653,7 @@ int main(int argc, char** argv)
 		"Request document file path")("prefix,p",
 		po::value(&filenamePrefix),
 		"Prefix to use for the generated C++ filenames")("namespace,n",
-		po::value(&clientNamespace),
+		po::value(&schemaNamespace),
 		"C++ sub-namespace for the generated types")("source-dir",
 		po::value(&sourceDir),
 		"Target path for the <prefix>Client.cpp source file")("header-dir",
@@ -2973,7 +1671,7 @@ int main(int argc, char** argv)
 
 		// If you specify any of these parameters, you must specify all three.
 		buildCustom = !schemaFileName.empty() || !requestFileName.empty() || !filenamePrefix.empty()
-			|| !clientNamespace.empty();
+			|| !schemaNamespace.empty();
 
 		if (buildCustom)
 		{
@@ -2989,7 +1687,7 @@ int main(int argc, char** argv)
 			{
 				throw po::required_option("prefix");
 			}
-			else if (clientNamespace.empty())
+			else if (schemaNamespace.empty())
 			{
 				throw po::required_option("namespace");
 			}
@@ -3015,13 +1713,14 @@ int main(int argc, char** argv)
 
 	try
 	{
-		const auto files = graphql::client::Generator(
-			{
-				graphql::client::GeneratorClient { std::move(schemaFileName),
-					std::move(requestFileName),
-					std::move(filenamePrefix),
-					std::move(clientNamespace) },
-				graphql::client::GeneratorPaths { std::move(headerDir), std::move(sourceDir) },
+		const auto files = graphql::generator::client::Generator(
+			graphql::generator::SchemaOptions { std::move(schemaFileName),
+				std::move(filenamePrefix),
+				std::move(schemaNamespace) },
+			graphql::generator::client::GeneratorOptions {
+				std::move(requestFileName),
+				graphql::generator::client::GeneratorPaths { std::move(headerDir),
+					std::move(sourceDir) },
 				verbose,
 				noIntrospection,
 			})
